+++ src/main/java/firok/tiths/common/Traits.java
@@ -54,6 +54,7 @@ public class Traits
 	public static final AbstractTrait steamy=new TraitSteamy(); // 气动
 	public static final AbstractTrait treasureDetecting=new TraitTreasureDetecting(); // 宝藏感知
 	public static final AbstractTrait creaky=new TraitCreaky(); // 喀嚓
+	public static final AbstractTrait decoying=new TraitDecoying(); // 诱食
 	public static final AbstractTrait undeadCalling=new TraitUndeadCalling(); // 亡灵呼唤
 	public static final AbstractTrait repressing=new TraitRepressing(); // 压制
 }